Feat: #2 SplashScreen 구성
