+++ WAITHER/ios/WAITHER/Images.xcassets/Contents.json
@@ -1,6 +1,6 @@
 {
   "info" : {
-    "version" : 1,
-    "author" : "expo"
+    "author" : "xcode",
+    "version" : 1
   }
 }
