--- WAITHER/ios/WAITHER/Images.xcassets/SplashScreen.imageset/Contents.json
@@ -1,21 +1,0 @@
-{
-  "images": [
-    {
-      "idiom": "universal",
-      "filename": "splashscreen.png",
-      "scale": "1x"
-    },
-    {
-      "idiom": "universal",
-      "scale": "2x"
-    },
-    {
-      "idiom": "universal",
-      "scale": "3x"
-    }
-  ],
-  "info": {
-    "version": 1,
-    "author": "expo"
-  }
-}
