+++ WAITHER/package.json
@@ -12,10 +12,13 @@
     "expo": "~50.0.13",
     "expo-status-bar": "~1.11.1",
     "react": "18.2.0",
-    "react-native": "0.73.5"
+    "react-native": "0.73.5",
+    "react-native-splash-screen": "^3.3.0",
+    "styled-components": "^6.1.8"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.0"
+    "@babel/core": "^7.20.0",
+    "typescript": "^5.4.2"
   },
   "private": true
 }
